[wk] / update / home_screen / home_screen 작업
